added high-score system and devSettings window
Added a high-score system that stores player names and score in a textfile. And added a developer settings window to test functions and debug the game. Currently the dev Settings window allows you to add high scores and display the list. 

TODO: The high score system needs to be able to sort the list and display the highest score at the top of the text file.