added coordinates system, improved UI
So this was a big commit. Added a coordinates system that takes user input and sanitizes the living daylights out of it to check it's a valid set of coordinates. Also improved the UI by adding letters and numbers to the grids as well as labels for things and of course, a button and textbox