Removed devSettings.vb, moved highscore system to form1.vb
Massive commit. Really not ideal. Removed devSettings.vb because having some of the game code in a different window really wasn't ideal. Moved the highscore system to form1.vb. Game now asks for a player name, keeps track of score and when the player or computer wins, adds they're name and score to highscores.txt