Renamed most variables to have type prefix
Renamed most variables to have a type prefix indicating what type of variable they are. It is easier to work with variables when you know what type they are. Except for the variables in the function boardToArray because I don't have the time to understand what that code does right now