Added ship, water and hit/miss icons
Added ship, water and hit/miss icons to the two boards to make it much easier to understand what is on each board. The UI had to be tweaked to accommodate the bigger characters on the boards. I'm still not 100% happy with it as the water character isn't quite as wide as the ships so slightly miss aligns the columns on the boards. Also, the column number labels are miss-aligned with the columns now which will be fixed later if I have the time