Created windows installer
Created a Windows installer using Microsoft Visual Studio Installer Projects 2022. The installer installs VB-Warships v0.1.0 to "C:\Program Files (x86)\dublowduck\VB-Warships" and adds a shortcut to the Start menu but not the desktop. The default logos are used, for now. The installer is x86 so should work on all versions of Windows that VB-Warships supports.